Cherry-pick `aa1ba84` in to `7.49.0` (#16126)
This PR cherry-picks
https://github.com/MetaMask/metamask-mobile/commit/aa1ba84636c9b51a222bc53b910a6901eae3fc6a

PR: https://github.com/MetaMask/metamask-mobile/pull/16117

---------
