+++ app/components/Views/confirmations/components/info/switch-account-type/switch-account-type.tsx
@@ -2,9 +2,9 @@ import React, { useEffect } from 'react';
 import { View } from 'react-native';
 
 import { useConfirmationMetricEvents } from '../../../hooks/metrics/useConfirmationMetricEvents';
-import AccountNetworkInfo from '../../rows/account-network-info-row';
+import AccountNetworkInfoRow from '../../rows/account-network-info-row';
 import AdvancedDetailsRow from '../../rows/transactions/advanced-details-row/advanced-details-row';
-import GasFeesDetails from '../../rows/transactions/gas-fee-details';
+import GasFeesDetailsRow from '../../rows/transactions/gas-fee-details-row';
 import SwitchAccountTypeInfoRow from '../../rows/switch-account-type-info-row';
 
 const SwitchAccountType = () => {
@@ -13,9 +13,9 @@ const SwitchAccountType = () => {
 
   return (
     <View>
-      <AccountNetworkInfo />
+      <AccountNetworkInfoRow />
       <SwitchAccountTypeInfoRow />
-      <GasFeesDetails />
+      <GasFeesDetailsRow />
       <AdvancedDetailsRow />
     </View>
   );
