+++ app/components/Views/confirmations/components/rows/transactions/from-to-row/index.ts
@@ -1,0 +1,1 @@
+export { default } from './from-to-row';
