--- app/components/Views/confirmations/components/rows/transactions/from-to/from-to.styles.ts
@@ -1,27 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styleSheet = () => StyleSheet.create({
-    container: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      paddingTop: 2,
-      paddingBottom: 5,
-      paddingHorizontal: 4,
-    },
-    nameContainer: {
-      flex: 1,
-      display: 'flex',
-      flexDirection: 'row',
-    },
-    leftNameContainer: {
-      justifyContent: 'flex-start',
-    },
-    rightNameContainer: {
-      justifyContent: 'flex-end',
-    },
-    iconContainer: {
-      paddingHorizontal: 8,
-    },
-  });
-
-export default styleSheet;
