--- app/components/Views/confirmations/components/rows/transactions/from-to/index.ts
@@ -1,1 +1,0 @@
-export { default } from './from-to';
