+++ app/components/Views/confirmations/components/rows/transactions/gas-fee-details-row/index.ts
@@ -1,0 +1,1 @@
+export { default } from './gas-fee-details-row';
