--- app/components/Views/confirmations/components/rows/transactions/gas-fee-details/index.ts
@@ -1,1 +1,0 @@
-export { default } from './gas-fee-details';
