+++ app/components/Views/confirmations/components/rows/transactions/network-and-origin-row/network-and-origin-row.styles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styleSheet = () =>
+  StyleSheet.create({
+    infoRowOverride: {
+      paddingBottom: 4,
+    },
+    networkRowContainer: {
+      flexDirection: 'row',
+      alignItems: 'center',
+    },
+    avatarNetwork: {
+      marginRight: 4,
+    },
+  });
+
+export default styleSheet;
